requirements.txt
pyPdf==1.13
pillow==2.4.0
